requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Text extraction and processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.1.0
docx2txt>=0.8

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# AI and LLM
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.40
langsmith>=0.0.87
openai>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0

# Vector databases
chromadb>=0.4.15
faiss-cpu>=1.7.4

# Database and storage
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
pandas>=2.1.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Web framework utilities
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
aiofiles>=23.2.0

# Logging and monitoring
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development tools
black>=23.9.1
flake8>=6.1.0
mypy>=1.6.0